refactor: pull download-dist out into utils
